Made the prometheus_session decorator do nothing if given a None directory
Conditionally wrapping or not wrapping code in a context is a bit
unpleasant, so let's just always wrap it in a context that doesn't always
do something!

+++ src/os2datascanner/utils/prometheus.py
@@ -1,4 +1,5 @@
 import os
+import os.path
 import json
 import errno
 from random import randrange
@@ -9,38 +10,42 @@ from prometheus_client import start_http_server
 
 @contextmanager
 def prometheus_session(name, advertisement_directory, **kwargs):
-    # Find a free port to serve Prometheus metrics over...
-    while True:
-        try:
-            # start_http_server tells us nothing about the HTTP server thread
-            # that it created, so we can't just pass a port of 0 -- we need to
-            # know which port it's using in order to be able to create the
-            # advertisement file (groan)
-            port = randrange(5000, 65000)
-            start_http_server(port)
-            break
-        except OSError as ex:
-            if ex.errno == errno.EADDRINUSE:
-                continue
-            else:
-                raise
+    if advertisement_directory:
+        # Find a free port to serve Prometheus metrics over...
+        while True:
+            try:
+                # start_http_server tells us nothing about the HTTP server
+                # thread that it created, so we can't just pass a port of 0 --
+                # we need to know which port it's using in order to be able to
+                # create the advertisement file (groan)
+                port = randrange(5000, 65000)
+                start_http_server(port)
+                break
+            except OSError as ex:
+                if ex.errno == errno.EADDRINUSE:
+                    continue
+                else:
+                    raise
 
-    # ... advertise this port, and this service, to Prometheus...
-    os.makedirs(advertisement_directory, exist_ok=True)
-    with NamedTemporaryFile(mode="wt", dir=advertisement_directory,
-            delete=False) as fp:
-        tmpfile = fp.name
-        # (... making sure that other users can read the advertisement...)
-        os.chmod(tmpfile, 0o644)
+        # ... advertise this port, and this service, to Prometheus...
+        os.makedirs(advertisement_directory, exist_ok=True)
+        with NamedTemporaryFile(mode="wt", dir=advertisement_directory,
+                delete=False) as fp:
+            tmpfile = fp.name
+            # (... making sure that other users can read the advertisement...)
+            os.chmod(tmpfile, 0o644)
 
-        json.dump([
-            {
-                "targets": ["localhost:{0}".format(port)],
-                "labels": kwargs
-            }
-        ], fp)
-    advertisement_path = advertisement_directory + ("/{0}.json".format(name))
-    os.rename(tmpfile, advertisement_path)
+            json.dump([
+                {
+                    "targets": ["localhost:{0}".format(port)],
+                    "labels": kwargs
+                }
+            ], fp)
+        advertisement_path = os.path.join(
+                advertisement_directory, "{0}.json".format(name))
+        os.rename(tmpfile, advertisement_path)
+    else:
+        advertisement_path = None
 
     try:
         # ... and yield to execute whatever's in the with block
@@ -48,4 +53,5 @@ def prometheus_session(name, advertisement_directory, **kwargs):
     finally:
         # Now that we're back and the context has finished, delete the service
         # advertisement
-        os.unlink(advertisement_path)
+        if advertisement_path:
+            os.unlink(advertisement_path)
